[Refactoring] Move code to cpp as far as CNF links with cocaine-core.
